+++ examples/react-client/videoroom/src/App.tsx
@@ -27,10 +27,20 @@ function App() {
 
       <div className="w-full h-full p-4">
         <section className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-4">
-          {localParticipant && <VideoTracks videoTracks={localParticipant.videoTracks} name="You" id="0" />}
+          {localParticipant && (
+            <VideoTracks
+              videoTracks={localParticipant.videoTracks}
+              name="You"
+              id="0"
+            />
+          )}
 
           {participants.map(({ id, videoTracks, metadata }) => (
-            <VideoTracks videoTracks={videoTracks} name={(metadata as { name?: string })?.name ?? id} id={id} />
+            <VideoTracks
+              videoTracks={videoTracks}
+              name={(metadata as { name?: string })?.name ?? id}
+              id={id}
+            />
           ))}
           {participants.map(({ id, audioTracks }) => (
             <AudioTracks audioTracks={audioTracks} key={id} />
