+++ examples/react-client/videoroom/src/components/AudioVisualizer.tsx
