+++ examples/react-client/videoroom/src/components/VideoPlayer.tsx
